+++ plugins/core/jetbrains-community/src/software/aws/toolkits/jetbrains/core/coroutines/scopes.kt
@@ -5,13 +5,11 @@ package software.aws.toolkits.jetbrains.core.coroutines
 
 import com.intellij.openapi.Disposable
 import com.intellij.openapi.application.Application
-import com.intellij.openapi.components.service
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.util.Disposer
-import kotlinx.coroutines.CoroutineName
 import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.SupervisorJob
 import kotlinx.coroutines.cancel
+import migration.software.aws.toolkits.jetbrains.core.coroutines.PluginCoroutineScopeTracker
 import java.util.concurrent.CancellationException
 
 /**
@@ -74,25 +72,3 @@ inline fun <reified T : Any> T.projectCoroutineScope(project: Project): Coroutin
 )
 inline fun <reified T : Any> T.disposableCoroutineScope(disposable: Disposable): CoroutineScope =
     disposableCoroutineScope(disposable, T::class.java.name)
-
-class PluginCoroutineScopeTracker : Disposable {
-    @PublishedApi
-    internal fun applicationThreadPoolScope(coroutineName: String): CoroutineScope = BackgroundThreadPoolScope(coroutineName, this)
-
-    override fun dispose() {}
-
-    companion object {
-        fun getInstance() = service<PluginCoroutineScopeTracker>()
-        fun getInstance(project: Project) = project.service<PluginCoroutineScopeTracker>()
-    }
-}
-
-private class BackgroundThreadPoolScope(coroutineName: String, disposable: Disposable) : CoroutineScope {
-    override val coroutineContext = SupervisorJob() + CoroutineName(coroutineName) + getCoroutineBgContext()
-
-    init {
-        Disposer.register(disposable) {
-            coroutineContext.cancel(CancellationException("Parent disposable was disposed"))
-        }
-    }
-}
